refactor: removed unwanted and duplicate codes + tested manually working fine
