+++ src/components/movie-view/movie-view.jsx
@@ -1,14 +1,11 @@
 import PropTypes from 'prop-types';
 import { Card, Button } from 'react-bootstrap';
 import { useParams } from 'react-router';
-import { BaseUrl } from "../../constants/constant";
+import { BaseUrl, loggedInUser, token } from "../../constants/constant";
 
 export const MovieView = ({ movies }) => {
     const { movieId } = useParams();
-    const loggedInUser = JSON.parse(localStorage.getItem('user'));
     const loggedInUsername = loggedInUser.userName;
-    const token = localStorage.getItem("token");
-
     const movie = movies.find((b) => b.id === movieId);
     const movieTitle = movie.title
 
@@ -26,25 +23,24 @@ export const MovieView = ({ movies }) => {
                 "Content-Type": "application/json"
             },
             body: JSON.stringify(requestBody)
-        })
-            .then((response) => {
-                if (response.status == 201) {
-                    alert("Movie added successfully to favorite list");
-                } else {
-                    alert("Unable to add list");
-                }
-            })
-            .catch((error) => {
-                console.error("Error:", error);
-            });
+        }).then((response) => {
+            if (response.status == 201) {
+                alert("Movie added successfully to favourite list.");
+            } else {
+                alert("Unable to add movie to favourite list.");
+            }
+        }).catch((error) => {
+            console.error("Error:", error);
+        });
     };
 
     return (
         <>
             <Card style={{ maxWidth: "50rem" }}>
-                <Card.Img className="img-fluid" src={movie.image} alt="movieImg" />
+                <Card.Img className="img-fluid" src={movie.image} alt="logoImg" />
             </Card>
-            <Button onClick={addFavHandler}>AddToFab</Button>
+
+            <Button className='m-2' variant='warning' onClick={addFavHandler}>Add To Favourite</Button>
             <div className='movieDescriptionHeader'><span>Title: </span>{movie.title}</div>
             <div className='movieDescriptionHeader'><span>Director: </span>{movie.director}</div>
             <div className='movieDescriptionHeader'>{movie.directorBio}</div>
