+++ src/components/navigation-bar/navigation-bar.jsx
@@ -1,6 +1,6 @@
 import { Navbar, Nav, Image, Container, Button } from "react-bootstrap"
 import logo from "../../css/logo.png"
-import { Link } from 'react-router-dom'
+import { size } from "lodash"
 
 export const NavigationBar = ({ user, onLoggedOut }) => {
 
@@ -9,14 +9,14 @@ export const NavigationBar = ({ user, onLoggedOut }) => {
             <>
                 {user && (
                     < Navbar bg="black" data-bs-theme="dark">
-                        <Navbar.Brand href="#"><Image src={logo} height={45} style={{ color: "white", marginLeft: "8px" }} /></Navbar.Brand>
+                        <Navbar.Brand href="##"><Image src={logo} alt="logoImg" height={45} style={{ color: "white", marginLeft: "8px" }} /></Navbar.Brand>
                         <Container>
                             <Navbar.Toggle aria-controls="basic-navbar-nav" />
                             <Navbar.Collapse id="basic-navbar-nav">
                                 <Nav className="me-auto">
-                                    <Nav.Link style={{ color: "white" }} href="/">Movies</Nav.Link>
+                                    <Nav.Link style={{ color: "white", fontSize: "18px" }} href="/">Movies</Nav.Link>
 
-                                    <Nav.Link style={{ color: "white" }} href="/profile">Profile</Nav.Link>
+                                    <Nav.Link style={{ color: "white", fontSize: "18px" }} href="/profile">Profile</Nav.Link>
                                 </Nav>
                                 <Nav.Item className="ml-auto">
                                     <Button className="float-right" style={{ background: "gray", border: 'none' }} size="sm" onClick={onLoggedOut}> Logout</Button>
@@ -37,8 +37,8 @@ export const NavigationBar = ({ user, onLoggedOut }) => {
                             <Navbar.Toggle aria-controls="basic-navbar-nav" />
                             <Navbar.Collapse id="basic-navbar-nav">
                                 <Nav className="me-auto">
-                                    <Nav.Link style={{ color: "white" }} href="/register">Register</Nav.Link>
-                                    <Nav.Link style={{ color: "white" }} href="/login">Login</Nav.Link>
+                                    <Nav.Link style={{ color: "white", fontSize: "18px" }} href="/register">Register</Nav.Link>
+                                    <Nav.Link style={{ color: "white", fontSize: "18px" }} href="/login">Login</Nav.Link>
                                 </Nav>
                             </Navbar.Collapse>
                         </Container>
