+++ src/constants/constant.jsx
@@ -1,1 +1,3 @@
 export const BaseUrl = "https://flix-api-1faf.onrender.com";
+export const loggedInUser = JSON.parse(localStorage.getItem('user'));
+export const token = localStorage.getItem("token");
