+++ src/index.scss
@@ -27,7 +27,7 @@ $color: rgb(250, 247, 247);
 
     color: #ffff;
     font-size: 18px;
-    margin: 8px;
+    margin:3px;
 }
 
 .signIn {
